Recursively assign threats instead of re-initializing
The old way of doing things was to assign threats and then check if
they abided by a set of rules, and if they didn't, restart everything
including the generation of threat levels.

The problem with this is that certain valid threat generations
are less likely to survive the rules, for example generations with lots
of threats. So these stood a smaller chance, statistically, of showing
up in games, even if they are perfectly valid generations.

This re-write does an exhaustive search for a valid distribution of
theats to timeslots and only re-generate if there is no distribution
of time slots that are valid against all rules.

And, I also added a second 'balance' rule, which is that the difference
in threat level between phase 1 and phase 2 can't be more than 2.
(I checked every mission, and this seemed to be a universal rule.)